refactor synchronization for simplicity
