+++ zarr/sync.py
@@ -18,22 +18,18 @@ class ThreadSynchronizer(object):
 
     def __init__(self):
         self.mutex = Lock()
-        self.attrs_lock = Lock()
-        self.chunk_locks = defaultdict(Lock)
+        self.locks = defaultdict(Lock)
 
-    def chunk_lock(self, ckey):
+    def __getitem__(self, item):
         with self.mutex:
-            lock = self.chunk_locks[ckey]
-        return lock
+            return self.locks[item]
 
     def __getstate__(self):
         return dict()
 
     def __setstate__(self, d):
         # reinitialize from scratch
-        self.mutex = Lock()
-        self.attrs_lock = Lock()
-        self.chunk_locks = defaultdict(Lock)
+        self.__init__()
 
 
 class ProcessSynchronizer(object):
@@ -51,104 +47,10 @@ class ProcessSynchronizer(object):
     def __init__(self, path):
         self.path = path
 
-    @property
-    def attrs_lock(self):
-        return fasteners.InterProcessLock(
-            os.path.join(self.path, 'attrs.lock')
-        )
-
-    def chunk_lock(self, ckey):
+    def __getitem__(self, item):
         lock = fasteners.InterProcessLock(
-            os.path.join(self.path, '%s.lock' % ckey)
+            os.path.join(self.path, '%s.lock' % item)
         )
         return lock
 
-
-class SynchronizedArray(Array):
-    """Instantiate a synchronized array.
-
-    Parameters
-    ----------
-    store : MutableMapping
-        Array store, already initialized.
-    synchronizer : object
-        Array synchronizer.
-    readonly : bool, optional
-        True if array should be protected against modification.
-    path : string, optional
-        Path under which array is stored.
-
-    Examples
-    --------
-    >>> import zarr
-    >>> store = dict()
-    >>> zarr.init_array(store, shape=1000, chunks=100)
-    >>> synchronizer = zarr.ThreadSynchronizer()
-    >>> z = zarr.SynchronizedArray(store, synchronizer)
-    >>> z
-    zarr.sync.SynchronizedArray((1000,), float64, chunks=(100,), order=C)
-      compression: blosc; compression_opts: {'clevel': 5, 'cname': 'lz4', 'shuffle': 1}
-      nbytes: 7.8K; nbytes_stored: 285; ratio: 28.1; initialized: 0/10
-      store: builtins.dict; synchronizer: zarr.sync.ThreadSynchronizer
-
-    Notes
-    -----
-    Only writing data to the array via the __setitem__() method and
-    modification of user attributes are synchronized. Neither append() nor
-    resize() are synchronized.
-
-    Writing to the array is synchronized at the chunk level. I.e.,
-    the array supports concurrent write operations via the __setitem__()
-    method, but these will only exclude each other if they both require
-    modification of the same chunk.
-
-    """  # flake8: noqa
-
-    def __init__(self, store, synchronizer, readonly=False, path=None,
-                 chunk_store=None):
-        super(SynchronizedArray, self).__init__(store, readonly=readonly,
-                                                path=path,
-                                                chunk_store=chunk_store)
-        self.synchronizer = synchronizer
-        akey = self._key_prefix + attrs_key
-        self._attrs = SynchronizedAttributes(store, synchronizer, key=akey,
-                                             readonly=readonly)
-
-    def _chunk_setitem(self, cidx, key, value):
-        ckey = self._ckey(cidx)
-        with self.synchronizer.chunk_lock(ckey):
-            super(SynchronizedArray, self)._chunk_setitem(cidx, key, value)
-
-    def __repr__(self):
-        r = super(SynchronizedArray, self).__repr__()
-        r += ('; synchronizer: %s.%s' %
-              (type(self.synchronizer).__module__,
-               type(self.synchronizer).__name__))
-        return r
-
-    def __getstate__(self):
-        return self._store, self.synchronizer, self._readonly, self._path, \
-               self._chunk_store
-
-    def __setstate__(self, state):
-        self.__init__(*state)
-
-
-class SynchronizedAttributes(Attributes):
-
-    def __init__(self, store, synchronizer, key='attrs', readonly=False):
-        super(SynchronizedAttributes, self).__init__(store, key=key,
-                                                     readonly=readonly)
-        self.synchronizer = synchronizer
-
-    def __setitem__(self, key, value):
-        with self.synchronizer.attrs_lock:
-            super(SynchronizedAttributes, self).__setitem__(key, value)
-
-    def __delitem__(self, key):
-        with self.synchronizer.attrs_lock:
-            super(SynchronizedAttributes, self).__delitem__(key)
-
-    def update(self, *args, **kwargs):
-        with self.synchronizer.attrs_lock:
-            super(SynchronizedAttributes, self).update(*args, **kwargs)
+    # pickling and unpickling should be handled automatically
